requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Booktype==1.5
certifi==2025.4.26
click==8.2.0
distro==1.9.0
Django==5.2.1
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
openai==1.79.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.1.0
setuptools==80.7.1
simplejson==3.20.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
